Assignment_2/flask_test/requirements.txt
flask
mysql-connector-python
Flask-SQLAlchemy
SQLAlchemy
pymysql